fix: remove decimals from days and add black text color
- Convert daysElapsed and daysRemaining to integers (no decimals)
- Add black color (#000) to days-remaining and days-elapsed displays
- Apply contract-info-text class to all contract details
- Add timeline-stat-value class for bold black text on stats

+++ app/Services/ContractTimelineService.php
@@ -16,9 +16,9 @@ class ContractTimelineService
         $now = now();
 
         // Cálculos
-        $daysElapsed = $startDate->diffInDays($now);
-        $totalDays = $startDate->diffInDays($endDate);
-        $daysRemaining = max(0, $now->diffInDays($endDate));
+        $daysElapsed = (int) $startDate->diffInDays($now);
+        $totalDays = (int) $startDate->diffInDays($endDate);
+        $daysRemaining = (int) max(0, $now->diffInDays($endDate));
         $progress = $totalDays > 0 ? round(($daysElapsed / $totalDays) * 100, 1) : 0;
 
         // Data sugerida para renovação (60 dias antes)
